# openapi.yml
openapi: 3.0.3
info:
  title: ASPB Autowebinar API
  version: 1.0.0
servers:
  - url: /
paths:
...
  /api/v1/auth/passwordless/request:
    post:
      summary: Request a passwordless platform User login link
      description: >-
        Returns the same accepted response for known, unknown, disabled and organization-less
        addresses. Client-supplied organization identifiers are rejected.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordlessLoginRequest'
      responses:
        '202':
          description: Generic non-enumerating response; a durable email job may have been queued
        '400':
          description: Strict request payload validation failed
        '403':
          description: CSRF token is missing or invalid
        '404':
          description: Platform accounts compatibility flag is disabled
  /api/v1/auth/passwordless/consume:
    post:
      summary: Consume one-time platform User login token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OneTimeTokenRequest'
      responses:
        '200':
          description: Hash-backed HttpOnly User session cookie issued; eligible memberships returned
        '400':
          description: Strict request payload validation failed
        '401':
          description: Token is unknown, expired, wrong-purpose or already consumed
        '403':
          description: CSRF token is missing or invalid
  /api/v1/auth/session:
    get:
      summary: Read the current platform User session
      security:
        - userSessionCookie: []
      responses:
        '200':
          description: User, active organization and server-resolved active memberships
        '401':
          description: User session is missing, expired, revoked or no longer eligible
  /api/v1/auth/active-organization:
    post:
      summary: Select one active organization from authenticated User memberships
      security:
        - userSessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActiveOrganizationRequest'
      responses:
        '200':
          description: Active organization selected in the server-side session
        '401':
          description: User session is missing or invalid
        '404':
          description: Unknown and foreign organizations are intentionally indistinguishable
  /api/v1/auth/logout:
    post:
      summary: Revoke the current platform User session
      security:
        - userSessionCookie: []
      responses:
        '204':
          description: Current session revoked and cookie cleared
  /api/v1/auth/sessions/revoke-all:
    post:
      summary: Revoke all sessions for the authenticated platform User
      security:
        - userSessionCookie: []
      responses:
        '204':
          description: Session version incremented, sessions revoked and cookie cleared
  /api/v1/auth/mfa/verify:
    post:
      summary: Verify an MFA-challenged platform User session
      description: Until this succeeds, the session cannot read tenant data or call protected tenant endpoints.
      security:
        - userSessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaOtpRequest'
      responses:
        '200':
          description: Current session marked MFA-verified and tenant session summary returned
        '400':
          description: Strict OTP payload validation failed
        '401':
          description: Session or one-time code is invalid
  /api/v1/auth/mfa/enrollment/start:
    post:
      summary: Start a ten-minute owner MFA enrollment
      security:
        - userSessionCookie: []
      responses:
        '200':
          description: One-time setup secret and otpauth URL returned with no-store caching
        '403':
          description: Active membership in the selected organization is not OWNER
        '409':
          description: MFA is already enabled
  /api/v1/auth/mfa/enrollment/confirm:
    post:
      summary: Confirm owner MFA enrollment
      security:
        - userSessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaOtpRequest'
      responses:
        '200':
          description: MFA enabled, current session verified and other User sessions revoked
        '401':
          description: One-time code is invalid
        '403':
          description: Active membership in the selected organization is not OWNER
        '409':
          description: Enrollment is absent, expired or MFA is already enabled
  /api/v1/auth/mfa/disable:
    post:
      summary: Disable owner MFA after a fresh TOTP check
      security:
        - userSessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaOtpRequest'
      responses:
        '200':
          description: MFA secret removed and other User sessions revoked
        '401':
          description: Current session or one-time code is invalid
        '403':
          description: Active membership in the selected organization is not OWNER
  /api/v1/organization/invitations/accept:
    post:
      summary: Accept a role-bound one-time organization invitation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OneTimeTokenRequest'
      responses:
        '200':
          description: Membership created or reactivated and HttpOnly User session issued
        '401':
          description: Invitation is unknown, expired, revoked or already consumed
  /api/v1/webinar-invitations/accept:
    post:
      summary: Accept an email-bound private Webinar invitation
      security:
        - userSessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OneTimeTokenRequest'
      responses:
        '200':
          description: Single-use token consumed and the authenticated verified User bound to the grant
        '400':
          description: Strict token payload validation failed
        '401':
          description: User session is missing, expired or MFA-challenged
        '404':
          description: Unknown, wrong-email, expired, revoked and already-used invitations are intentionally indistinguishable
  /api/v1/organizations:
    post:
      summary: Create the first organization for an authenticated verified User
      description: The server rejects Users who already have an active membership. Organization, active HUMAN OWNER membership, session tenant selection, audit and idempotency record commit atomically. This never creates a platform-admin identity.
      security:
        - userSessionCookie: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, minLength: 8, maxLength: 128, pattern: '^[A-Za-z0-9._:-]+$' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrganizationCreateRequest' }
      responses:
        '201': { description: Organization and OWNER membership created atomically }
        '200': { description: Same successful result replayed for the same idempotency key and payload }
        '400': { description: Strict payload, slug or idempotency contract failed }
        '401': { description: User session is missing, expired or MFA-challenged }
        '409': { description: Safe create conflict or idempotency key payload conflict }
        '429': { description: Production organization-creation rate limit exceeded }
  /api/v1/organizations/{organizationId}:
    parameters:
      - name: organizationId
        in: path
        required: true
        schema: { type: string, minLength: 1, maxLength: 191 }
    get:
      summary: Read safe settings for the server-selected organization
      security:
        - userSessionCookie: []
      responses:
        '200': { description: Owner-visible organization settings projection; private no-store response }
        '403': { description: Active membership is not OWNER }
        '404': { description: Unknown and cross-tenant organizations are intentionally indistinguishable }
    patch:
      summary: Update safe organization settings with optimistic revision and idempotency
      security:
        - userSessionCookie: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, minLength: 8, maxLength: 128, pattern: '^[A-Za-z0-9._:-]+$' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrganizationUpdateRequest' }
      responses:
        '200': { description: Settings updated and audited, or the exact idempotent response replayed }
        '400': { description: Strict payload, timezone or idempotency contract failed }
        '403': { description: Active membership is not OWNER }
        '404': { description: Unknown and cross-tenant organizations are intentionally indistinguishable }
        '409': { description: Revision conflict or idempotency key payload conflict }
  /api/v1/organizations/{organizationId}/members:
    get:
      summary: List members in the server-selected organization with cursor pagination
      security:
        - userSessionCookie: []
      parameters:
        - name: organizationId
          in: path
          required: true
          schema: { type: string, minLength: 1, maxLength: 191 }
        - name: cursor
          in: query
          schema: { type: string, minLength: 1, maxLength: 191 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 25 }
      responses:
        '200': { description: Minimal access-management member projection and an optional nextCursor }
        '403': { description: Active membership is not OWNER }
        '404': { description: Unknown and cross-tenant organizations are intentionally indistinguishable }
  /api/v1/organization/invitations:
    post:
      summary: Create an invitation in the server-selected organization
      security:
        - userSessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationInvitationRequest'
      responses:
        '201':
          description: Pending invitation and durable delivery job created atomically
        '403':
          description: Authenticated membership is not an active owner
        '409':
          description: The email already has an active membership
    get:
      summary: List invitations in the server-selected organization
      security:
        - userSessionCookie: []
      parameters:
        - name: cursor
          in: query
          schema: { type: string, minLength: 1, maxLength: 191 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 25 }
      responses:
        '200':
          description: Tenant-scoped invitation list; expired pending rows are marked expired
        '403':
          description: Authenticated membership is not an active owner
  /api/v1/organization/invitations/{invitationId}:
    delete:
      summary: Revoke a pending invitation in the server-selected organization
      security:
        - userSessionCookie: []
      parameters:
        - $ref: '#/components/parameters/InvitationId'
      responses:
        '200':
          description: Invitation revoked, active token hashes invalidated and delivery cancelled when possible
        '403':
          description: Authenticated membership is not an active owner
        '404':
          description: Unknown and cross-tenant invitations are intentionally indistinguishable
  /api/v1/organization/memberships/{membershipId}/role:
    patch:
      summary: Change a membership role in the server-selected tenant
      security:
        - userSessionCookie: []
      parameters:
        - $ref: '#/components/parameters/MembershipId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MembershipRoleRequest'
      responses:
        '200':
          description: Membership updated and audited atomically
        '403':
          description: Authenticated membership is not an active owner
        '404':
          description: Unknown and cross-tenant memberships are intentionally indistinguishable
        '409':
          description: Mutation would remove the last available active human owner
  /api/v1/organization/memberships/{membershipId}:
    delete:
      summary: Remove a membership from the server-selected tenant
      security:
        - userSessionCookie: []
      parameters:
        - $ref: '#/components/parameters/MembershipId'
      responses:
        '200':
          description: Membership removed and audited atomically
        '403':
          description: Authenticated membership is not an active owner
        '404':
          description: Unknown and cross-tenant memberships are intentionally indistinguishable
        '409':
          description: Mutation would remove the last available active human owner
...
  /api/admin/platform/organizations/{id}:
    patch:
      summary: Change platform organization state with rollback evidence
      security: [{ adminSessionCookie: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string, maxLength: 191 } }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PlatformOrganizationChange' } } }
      responses:
        '200': { description: Owner-only confirmed organization change committed and audited }
        '409': { description: Optimistic revision conflict }
...
  /api/v1/organizations/{organizationId}/retention/plan:
    post:
      summary: Produce a non-mutating tenant retention inventory plan
      description: OWNER/AUDITOR only. Cutoffs stay empty until policy terms are approved; active legal holds are reflected by category.
      security: [{ userSessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [confirmDryRun]
              properties: { confirmDryRun: { type: boolean, enum: [true] } }
      responses:
        '200': { description: Deterministic dry-run plan; no rows changed }
        '403': { description: Active membership is not OWNER or AUDITOR }
        '404': { description: Unknown or cross-tenant organization }
  /api/v1/organizations/{organizationId}/retention/apply:
    post:
      summary: Reject destructive retention apply while terms remain unapproved
      security: [{ userSessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '404': { description: Unknown or cross-tenant organization }
        '409': { description: Always returns retention_apply_blocked in this release }
...
components:
  parameters:
...
    MembershipId:
      name: membershipId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 191
    InvitationId:
      name: invitationId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 191
...
    CsrfToken:
      name: x-csrf-token
      in: header
      required: true
      schema: { type: string, minLength: 1, maxLength: 4096 }
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 191 }
...
  schemas:
...
    PlatformOrganizationChange:
      type: object
      additionalProperties: false
      required: [expectedRevision, reason, confirmation]
      properties:
        name: { type: string, minLength: 2, maxLength: 191 }
        status: { type: string, enum: [ACTIVE, SUSPENDED, ARCHIVED] }
        expectedRevision: { type: integer, minimum: 1 }
        reason: { type: string, minLength: 3, maxLength: 500 }
        confirmation: { type: string, enum: [CONFIRM_PLATFORM_CHANGE] }
...
    OneTimeTokenRequest:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token:
          type: string
          minLength: 43
          maxLength: 43
          pattern: '^[A-Za-z0-9_-]{43}$'
    PasswordlessLoginRequest:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email:
          type: string
          format: email
          maxLength: 320
    ActiveOrganizationRequest:
      type: object
      additionalProperties: false
      required: [organizationId]
      properties:
        organizationId:
          type: string
          minLength: 1
          maxLength: 191
    OrganizationSettings:
      type: object
      additionalProperties: false
      properties:
        defaultTimezone:
          type: string
          minLength: 3
          maxLength: 64
          description: Valid IANA timezone name
        locale:
          type: string
          enum: [ru-RU]
    OrganizationCreateRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, minLength: 2, maxLength: 160 }
        slug:
          type: string
          minLength: 2
          maxLength: 63
          description: Normalized by the server; reserved and empty normalized forms are rejected without disclosing other tenants.
        settings: { $ref: '#/components/schemas/OrganizationSettings' }
    OrganizationUpdateRequest:
      type: object
      additionalProperties: false
      minProperties: 2
      required: [expectedRevision]
      properties:
        expectedRevision: { type: integer, minimum: 1 }
        name: { type: string, minLength: 2, maxLength: 160 }
        settings: { $ref: '#/components/schemas/OrganizationSettings' }
    MembershipRoleRequest:
      type: object
      additionalProperties: false
      required: [role]
      properties:
        role:
          type: string
          enum: [OWNER, AUTHOR, MODERATOR, CRM_MANAGER, ANALYST, AUDITOR]
    OrganizationInvitationRequest:
      type: object
      additionalProperties: false
      required: [email, role]
      properties:
        email:
          type: string
          format: email
          maxLength: 320
        role:
          type: string
          enum: [OWNER, AUTHOR, MODERATOR, CRM_MANAGER, ANALYST, AUDITOR]
    MfaOtpRequest:
      type: object
      additionalProperties: false
      required: [otp]
      properties:
        otp:
          type: string
          pattern: '^\d{6}$'
...
  securitySchemes:
    metricsBearer:
      type: http
      scheme: bearer
    userSessionCookie:
      type: apiKey
      in: cookie
      name: aspb_user_session
    participantSessionCookie:
      type: apiKey
      in: cookie
      name: aspb_room_token
    adminSessionCookie:
      type: apiKey
      in: cookie
      name: aspb_admin_session